+++ jgclark.Dashboard/src/react/components/Header/Header.jsx
@@ -289,7 +289,7 @@ const Header = ({ lastFullRefresh }: Props): React$Node => {
   // Note: this is a hack on iOS and iPadOS in modal mode, to allow the modal close button to be visible
   // FIXME: not working yet on iOS and iPadOS, and not logging either!
   const isModal = (pluginData.platform === 'iOS' || pluginData.platform === 'iPadOS') && !pluginData.mainWindowModeSupported
-  logInfo('Header', `isModal:${String(isModal)}, mainWindowModeSupported:${String(pluginData.mainWindowModeSupported)} on platform ${pluginData.platform}`)
+  logDebug('Header', `isModal:${String(isModal)}, mainWindowModeSupported:${String(pluginData.mainWindowModeSupported)} on platform ${pluginData.platform}`)
 
   // ----------------------------------------------------------------------
   // Render