+++ jgclark.Dashboard/CHANGELOG.md
@@ -9,6 +9,13 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - TODO: fix isNoteFromAllowedFolder() for teamspace or possibly 2025-W21.md
 -->
 
+## [2.4.0.b15] 2026-01-???
+### Changed
+- ??? fix displaying folder names
+- ??? dev: check winowID throughout
+- add note's icon and iconColor to display of note title (if set in note's frontmatter)
+- ??? do same for Projects section
+
 ## [2.4.0.b14] 2026-01-16
 ### Changed
 - force iPad and iPhone to run in "main window" mode, not "floating window" mode, if it can.  Where it does, then change layout of right-hand-side of Header.